Split test-features into part thats under feature and one that's not
